Move config logic to separate module
