+++ mealpy/__main__.py
@@ -1,0 +1,5 @@
+from mealpy.mealpy import cli
+
+
+if __name__ == '__main__':
+    cli()
